first and foremost, my goals was to lazy load this app as much as I can. at this time, I believe this is a performant choice.
here I tried to create just the homepage module
but i just went all in and tried to separate most
of what i need to do..

first of all, i needed a shared module that can be used
between modules. this way i can still do lazy loading, but
still build reusable code

while creating routes for my homepage, i took this time to setup
a pagenotfound error page, then utilized my shared module instead of
creating it under the app

the app will not be minimal as possible, this serves as the backbone
of the application. page not found shouldn't be a part of the backbone,
form me its just an addon to for user experience.

lastly, I wanted to explore creating navlinks but i don't want to wire
it up statically, so i used the router service, pulling the routes i created
earlier and I can now use that to dynamically generate Top level
links to my navigation